refactor connect stories
