+++ ux/react/src/Connect/demo/request-ethereum-signature.js
@@ -1,0 +1,10 @@
+/* eslint-disable react/jsx-filename-extension */
+import Typography from '@mui/material/Typography';
+
+export default function Demo() {
+  return (
+    <Typography variant="p" gutterBottom>
+      TODO
+    </Typography>
+  );
+}
